Delete functionality added to Task and TaskList

See merge request cse1105/2022-2023/teams/oopp-team-23!66

+++ client/src/main/java/client/scenes/BoardOverviewCtrl.java
@@ -69,7 +69,7 @@ public class BoardOverviewCtrl implements Initializable {
             Platform.runLater(() -> {
                 MFXButton button = new MFXButton((String) board.get(1));
                 button.setOnAction(event
-                        -> switchSceneToBoard(server.getBoardById((long) board.get(0))));
+                        -> switchSceneToBoard(server.getBoardById((long)(int) board.get(0))));
                 boards.getItems().add(button);
                 boardTitle.clear();
             });
